# docs/swagger.yaml
definitions:
...
  domain.Paciente:
    properties:
      apellido:
        type: string
      dni:
        type: string
      domicilio:
        type: string
      fechaAlta:
        type: string
      id:
        type: integer
      nombre:
        type: string
    required:
    - apellido
    - dni
    - domicilio
    - fechaAlta
    - nombre
    type: object
...
  handler.RequestPaciente:
    properties:
      apellido:
        type: string
      dni:
        type: string
      domicilio:
        type: string
      fechaAlta:
        type: string
      nombre:
        type: string
    type: object
...
  web.errorResponse:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
info:
  contact:
    name: Alejandra Marin - Vanesa Vilte
    url: https://github.com/AlejaMarin/Desafio-2-Go
  description: Implementación de una API que permite administrar la reserva de turnos
    para una clínica odontológica.
  title: SISTEMA DE RESERVA DE TURNOS
  version: "1.0"
paths:
...
  /pacientes:
    post:
      consumes:
      - application/json
      description: Crear paciente
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Body patient
        in: body
        name: paciente
        required: true
        schema:
          $ref: '#/definitions/domain.Paciente'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Paciente'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Agregar paciente
      tags:
      - Pacientes
  /pacientes/{id}:
    delete:
      description: Eliminar un paciente por ID
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: token
        in: header
        name: token
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Eliminar paciente
      tags:
      - Pacientes
    get:
      description: Obtener paciente por ID
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Paciente'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Obtener paciente
      tags:
      - Pacientes
    patch:
      consumes:
      - application/json
      description: Actualizar un paciente por alguno de sus campos
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RequestPaciente'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Paciente'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Actualizar paciente
      tags:
      - Pacientes
    put:
      consumes:
      - application/json
      description: Actualizar paciente por ID
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Body patient
        in: body
        name: dentista
        required: true
        schema:
          $ref: '#/definitions/domain.Paciente'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Paciente'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Actualizar paciente
      tags:
      - Pacientes
...
schemes:
- http
swagger: "2.0"